+++ amrviz/amrviz/src/Sensors.cpp
@@ -69,8 +69,8 @@ namespace amrviz
             
 
             RFID_sub = node->create_subscription<std_msgs::msg::String>(slash_str + selected_robot + RFID_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::RFID_update_cb, this, _1)); 
-            IR_sub_L = node->create_subscription<std_msgs::msg::Float32>(slash_str + selected_robot + LEFT_IR_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::IR_left_update_cb, this, _1)); 
-            IR_sub_R = node->create_subscription<std_msgs::msg::Float32>(slash_str + selected_robot + RIGHT_IR_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::IR_right_update_cb, this, _1)); 
+            IR_sub_L = node->create_subscription<std_msgs::msg::UInt16>(slash_str + selected_robot + LEFT_IR_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::IR_left_update_cb, this, _1)); 
+            IR_sub_R = node->create_subscription<std_msgs::msg::UInt16>(slash_str + selected_robot + RIGHT_IR_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::IR_right_update_cb, this, _1)); 
             encoder_sub_L = node->create_subscription<std_msgs::msg::Float32>(slash_str + selected_robot + LEFT_ENCODER_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::encoder_left_update_cb, this, _1)); 
             encoder_sub_R = node->create_subscription<std_msgs::msg::Float32>(slash_str + selected_robot + RIGHT_ENCODER_TOPIC, rclcpp::SystemDefaultsQoS(), std::bind(&Sensors::encoder_right_update_cb, this, _1)); 
         }    
@@ -86,12 +86,12 @@ namespace amrviz
         uiPanel->rfid_time_block->setText(QString::fromStdString(std::to_string(get_ros2_time())));
     }
 
-    void Sensors::IR_left_update_cb(const std_msgs::msg::Float32& msg){
+    void Sensors::IR_left_update_cb(const std_msgs::msg::UInt16& msg){
         //update the RFID data out
         uiPanel->IR_left_block->setText(QString::fromStdString(std::to_string(msg.data)));
     }
 
-    void Sensors::IR_right_update_cb(const std_msgs::msg::Float32& msg){
+    void Sensors::IR_right_update_cb(const std_msgs::msg::UInt16& msg){
         //update the RFID data out
         uiPanel->IR_right_block->setText(QString::fromStdString(std::to_string(msg.data)));
     }
